+++ app/components/Accordion.tsx
@@ -4,7 +4,7 @@ import {
   type AccordionItemProps,
 } from "~/components/AccordionItem";
 import { translations } from "~/services/translations/translations";
-import { useShouldPrint } from "~/services/hooks/useShouldPrint";
+import { useShouldPrint } from "./hooks/useShouldPrint";
 
 export type AccordionProps = Readonly<{
   items: AccordionItemProps[];
