+++ app/components/filesUpload/FilesUpload.tsx
@@ -4,13 +4,13 @@ import { useActionData } from "react-router";
 import { type ErrorMessageProps } from "~/components/inputs";
 import InputError from "~/components/inputs/InputError";
 import { type UserData } from "~/domains/userData";
-import { useJsAvailable } from "~/services/hooks/useJsAvailable";
 import {
   errorStyling,
   fileUploadLimit,
   type PDFFileMetadata,
 } from "~/util/file/pdfFileSchema";
 import { FilesUploadHeader } from "./FilesUploadHeader";
+import { useJsAvailable } from "../hooks/useJsAvailable";
 import { InlineNotice, type InlineNoticeProps } from "../InlineNotice";
 import { FileInput } from "../inputs/FileInput";
