+++ app/components/hooks/useJsAvailable.ts
