+++ app/components/hooks/useShouldPrint.ts
